requirements.txt
pyttsx3
SpeechRecognition
vosk
pyqt5
transformers
datasets
sounddevice
spacy
dateparser
apscheduler
boto3
cryptography
moviepy
Pillow
python-vlc

